Adjust dom structure to seperate functionality
